Only iterate over command buffers that need to be reset in pool reset
By default command pool reset iterates over all of the pool's command
buffers. In turn, command buffer reset checks whether the buffer was
begun and only resets when it was. For command pools with a large number
of unused command buffers this causes a performance hit by iterating
over extra objects and checking a flag in each one.

This change adds a command buffer set to the command pool that stores
all the buffers that had begin called on them. During the pool reset
only those buffers are processed.